sdhci: improve recovery behavior when SDCC is jammed
Ported from (CR) on K3.0.
Also contains (CR).

Some cards can make the SDC controller hang for no apparent reason.
There is a catch-all timer in the host driver to recover from this.
The elaborate retry logic in the MMC driver will cause long delays
as it repeatedly polls the card status through a dead controller.
The real timeout for status commands is deterministic and very
short, so there is no need to wait 8 seconds for the SDCC on CMD13.

Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>