msm: ipa: prevent parallel execution of pipe setup functions in SSR
In a SSR scenario, it is possible that ipa_setup_sys_pipe and
ipa_teardown_sys_pipe functions will be called simultaneously for the
IPA WAN DL pipe and cause potential issues.
A guard has been added to prevent this parallelism. In addition,
in case that SSR is currently in progress, the call to ipa_setup_sys_pipe
from the matching RMNET_IOCTL_SET_INGRESS_DATA_FORMAT IOCTL is not allowed.

Change-Id: I85a4dc8b2c3ac44ce54202193a377c8c455ae350
Acked-by: David Arinzon <darinzon@qti.qualcomm.com>
Signed-off-by: Sivan Reinstein <sivanr@codeaurora.org>