mmc: cmdq: add clock scaling support for CQ mode
Add clock scaling support for command queuing mode.
In CQ mode we count busy time from sending the first data request
until the completion of all the pending data requests.
This is done by starting the busy time on the first data request,
stopping it at the end of each data request and restarting it as
long as there are pending data requests.

The need to scale up/down the clocks is checked before sending each
request. Scaling the clocks up and down is done immediately in the
cmdqd context.
Before scaling the clocks the queue has to be drained and halted,
to allow sending the switch and tuning commands.

Change-Id: Ie856284261afc9a7f02489d2c9ab6a438e609f57
Signed-off-by: Maya Erez <merez@codeaurora.org>
[Resolved trivial merge conflicts]
Signed-off-by: Andrey Cherepkov <andreiskull@gmail.com>