USB: phy-msm-usb: Don't exit from low power mode during PM resume
With micro-A cable connected in host mode and upon system resume, USB is
exited from low power mode even if the wakeup source is not USB and enters
into low power mode again because there is nothing pending on USB side
to be done. This is unnecessary activity as there is nothing to be done at
USB side in pm resume. Whenever the wakeup is because of USB activity e.g.
micro-A cable connect or disconnect then this activity is handled by first
resuming the USB in the OTG driver.

Do not bring USB out of low power mode, we don't need to perform root hub
or port resume in ehci pm resume. Make this function nop and let the
OTG driver handle this where DP and DM line change can be taken care by
the aync_irq and phy_irq bring USB out of low power mode in case of host
cable disconnect and connect.

Enable both async and core irq when USB is entering into low power mode
and set both the irqs wakeup capable. Async irq is now mapped to DP and
core irq is mapped to DM lines. Hence when it is host bus suspend case,
DP or DM line transition can wakeup the target if target is in deep sleep
state.

Change-Id: If8e0584c12d4cb7273992a56c19778fe814b3404
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>