[fixed] cpuidle: Add nop idle governor.
The lpm cpuidle driver does all its idle state calculations
in the driver, as opposed to in the governor. To avoid pointless
work, add a null cpuidle governor that returns fixed idle
state to move to, the result of which can be ignored.

Change-Id: I7e29513126e91023af13794792235bc4b5500d06
Signed-off-by: Riley Andrews <riandrews@google.com>

Conflicts:
	drivers/cpuidle/Kconfig

no rework of cpuidle ... YET ... in this branch

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>