msm: ipa: fix to system suspend
When there are no wakelocks acquired, system tries to go to
suspend. And if any of IPA pipes are in polling state,
system is not allowed go to suspend by IPA driver. As a result
system resumes its operation from suspend to full state and this
leads to bad power gains. Use a wakelock when any of the pipes
are in polling state to not to make the system go to suspend in
the first place. Release the wakelock when switching to interrupt mode.

Change-Id: I0dbfe5c65dceeb487955b130788890dec71a7efd
Acked-by: Chaitanya Pratapa <cpratapa@qti.qualcomm.com>
Signed-off-by: Ravinder Konka <rkonka@codeaurora.org>
Reviewed-on: http://gerrit.mot.com/791359
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: David Ding <dding@motorola.com>
Reviewed-by: Yi-Wei Zhao <gbjc64@motorola.com>