USB: phy-msm-usb: Don't enable usb irqs wakeup capability during disconnect
In PM suspend and resume with DCP charger connected case, Some times
core irq and async irqs are getting fired and usb exiting from low
power mode. As we are not scheduling the state machine work for charger
connect/disconnect case. We exited from the low power mode but after
that it is not entering low power mode again.

In the current code we are enabling both async and core irq when
USB is entering into low power mode and set both the irqs wakeup capable.
Async irq is now mapped to DP and core irq is mapped to DM lines. By this
any DP or DM line transition can wakeup the target if target is in deep
sleep state. To avoid this enable the wakeup capability for usb irqs only
in host bus suspend/device bus suspend case in LPM sequence.

Change-Id: I868cf319de9bf63069754b8e91a7cae58a1cab7c
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>