sched: fair: Return current CPU if search_cpu(s) is empty
If a task is affined to CPUs that are offline, select_best_cpu
and best_small_task_cpu will attempt to select a cpu from an
empty mask, resulting in corruption/crashes. Fix this by
returning the current CPU if search_cpu(s_ is empty.

Change-Id: Ib56a8a2513b0ae7ec8414b5ed3134e88b24dd2ac
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>

Conflicts:
	kernel/sched/fair.c

properly add 'sched: Return fallback CPU when no CPUs are available.':

sched: Return fallback CPU when no CPUs are available.
Given certain combinations of hotplug and cpusets, it is possible
to have a cpumask with no valid CPUs, including the CPU where the
task is currently running.

Fixes a regression caused by 93e62b823, because we will run the
body of the loop with an invalid CPU index instead of falling
through to return fallback_cpu.

bug 23748688

Change-Id: I621e4f6ab12f9fe932dfd3cd12d705d5617836bc

Tim Murray

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>