drivers:power: Fix the soc update delay
The soc value is the fg controller is 0-255. Reported soc is 0-100.
Current delta-soc = 255/100 = 3. That will skip some soc point.

I do have my doubts about this patch, but the charger code is spaghetti
and a bit hard to understand what does what and when.

Change-Id: I72b62fc8943322f31da6220e7401d6d9854314f7

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

Since this is hardly used anywhere with CONFIG_QPNP_FG_EXTENSION it's a
NOOP change but adding here anyway in case we'll need it later

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>