msm: mdss: control DSI clocks efficiently when sending DCS commands
In the current implementation it is possible that the DSI link clocks
could be toggled multiple times when sending a single DCS command. This
can result in significant delays leading to a poor user experience.
Fix this by ensuring that the clocks are enabled at the top level
function and any subsequent calls to enable the clocks would just result
in a ref count increase.

Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
Signed-off-by: Naseer Ahmed <naseer@codeaurora.org>