sched: Return fallback CPU when no CPUs are available.
Given certain combinations of hotplug and cpusets, it is possible
to have a cpumask with no valid CPUs, including the CPU where the
task is currently running.

Fixes a regression caused by 93e62b823, because we will run the
body of the loop with an invalid CPU index instead of falling
through to return fallback_cpu.

bug 23748688

Change-Id: I621e4f6ab12f9fe932dfd3cd12d705d5617836bc

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>