msm: ipa: Fix wakelock issue
There is a possibility that wakelock is not acquired when
polling state is set and we allow system to go to suspend.
Make a change to acquire/release the wakelock based on the
polling state to not allow system to go to suspend.

Bug: 24338185
Change-Id: Iec4e7511e9a03b4f9911b280ce4ddc0e913b5185
Acked-by: Chaitanya Pratapa <cpratapa@qti.qualcomm.com>
Signed-off-by: Ravinder Konka <rkonka@codeaurora.org>
Signed-off-by: Niranjan Pendharkar <npendhar@codeaurora.org>

Conflicts:
	drivers/platform/msm/ipa/ipa_i.h

due to
'msm: ipa: add support to get tether stats'
and
'msm: ipa: poll prod pipes when irq srcs is null'

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>