ASoC: wcd_cpe: Make sure messages are not dropped
It is possible that the client may send message, but the thread
processing messages will go back to wait state without processing the
message. In such case, the message will timeout and will never be
transmitted. Handle this case by making sure the worker thread is
synchronized with respect to the client thread that is sending the
message.

Bug: 26450651

Change-Id: Ic667b27d5c2a4469879327b90471bbc5464065fd
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Mekala Natarajan <mekalan@codeaurora.org>