cpufreq: interactive: prevents the frequency to directly raise above the
hispeed_freq from a lower frequency.

When the load was below
go_hispeed_load, there is a possibility that
choose_freq() would return
a frequency which would be higher than the
hispeed_freq. According to
the policy we should first jump to the
hispeed_freq, stay there for
above_hispeed_delay and then be allowed to
raise higher than
that.

Added a check to prevent the frequency to be directly raised
to
something higher than the hispeed_freq.

Change-Id:
Icda5d848dd9beadcc18835082ddf269732c75bd0
Signed-off-by: Ruchi Kandoi
<kandoiruchi@google.com>
(cherry picked from commit
4c7894cb01c7544225d5fcfb0fcb1d11fd2d1777)