fs/buffer.c: increase the buffer-head per-CPU LRU size
Increase the buffer-head per-CPU LRU size to allow efficient
filesystem
operations that access many blocks for each transaction.  For
example,
creating a file in a large ext4 directory with quota enabled
will access
multiple buffer heads and will overflow the LRU at the
default 8-block LRU
size:

* parent directory inode table block (ctime,
nlinks for subdirs)
* new inode bitmap
* inode table block
* 2 quota
blocks
* directory leaf block (not reused, but pollutes one cache
entry)
* 2 levels htree blocks (only one is reused, other pollutes
cache)
* 2 levels indirect/index blocks (only one is reused)

The
buffer-head per-CPU LRU size is raised to 16, as it shows in
metadata
performance benchmarks up to 10% gain for create, 4% for lookup
and 7% for
destroy.

Signed-off-by: Liang Zhen
<liang.zhen@intel.com>
Signed-off-by: Andreas Dilger
<andreas.dilger@intel.com>
Signed-off-by: Sebastien Buisson
<sebastien.buisson@bull.net>
Signed-off-by: Andrew Morton
<akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds
<torvalds@linux-foundation.org>
Signed-off-by: Pranav Vashi
<neobuddy89@gmail.com>