[PORT] USB: dwc3: msm: Kill tasklet before disabling irq in suspend
If tasklet dwc->bh and dwc3_msm_suspend() race each other, there is
chance that clocks are turned off as part of suspend() and end up
accessing registers as part of bottom half handler. This results in
crash. Fix this by killing tasklet before disabling irq as part of
suspend() so that tasklet gets completed if it is pending.

Change-Id: I63775dd27fedc0026d6edc6a76ec93e4581be644
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>

Conflicts:
	drivers/usb/dwc3/gadget.c

due to incorrect merge/position caused by inverse
order of occurrence of the functions

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>