PHYLIB: queue work on system_power_efficient_wq Phylib uses workqueues for multiple purposes. There is no real dependency of scheduling these on the cpu which scheduled them.
On a idle system, it is observed that and idle cpu wakes up many times just to
service this work. It would be better if we can schedule it on a cpu which the
scheduler believes to be the most appropriate one.

This patch replaces system_wq with system_power_efficient_wq for PHYLIB.

Cc: David S. Miller <davem at davemloft.net>
Cc: netdev at vger.kernel.org
Signed-off-by: Viresh Kumar <viresh.kumar at linaro.org>
Acked-by: David S. Miller <davem at davemloft.net>
Signed-off-by: Tejun Heo <tj at kernel.org>
(cherry picked from commit bbb47bdeae756f04b896b55b51f230f3eb21f207)
Signed-off-by: Mark Brown <broonie at linaro.org>