msm: mdss: use dedicated thread for blit request processing
Replace the PPP workqueue with a dedicated thread

The PPP driver has little work to do but it is critical that it be
completed on time. With a dedicated thread we can use the realtime
scheduler to achieve the desired performance characteristics.

This is modeled after the behavior of the display thread in the
mdss_fb driver.

Change-Id: I2a707f720875f19b029a03e25c62de8c896e666e
Git-commit: 7c0ed68cefdf72b1b74f59984057fff95dab52dc
Git-repo: https://android.googlesource.com/kernel/msm.git
Signed-off-by: Sachin Bhayare <sachin.bhayare@codeaurora.org>