fs:fuse: Disable shortcircuit when mmap is called on a file
When some data is written to a file both mmap and regular io
there can be race conditions that can cause incorrect data
to be saved.

Disable shortcircuit on the specific files on which  mmap is called
until we add mmap support to shortcircuit.

Bug: 30222859
Change-Id: I37626a9599f0cebd6238351dcdeff2392e4d9794
Signed-off-by: Nikhilesh Reddy <reddyn@codeaurora.org>

Conflicts:
	fs/fuse/file.c

remove comment about

'Use iocb->ki_pos instead of pos to handle the cases of files that are
opened with O_APPEND.'

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>