Not cool to only partially merge commits:
sched: fair: Return current CPU if search_cpu(s) is empty

If a task is affined to CPUs that are offline, select_best_cpu
and best_small_task_cpu will attempt to select a cpu from an
empty mask, resulting in corruption/crashes. Fix this by
returning the current CPU if search_cpu(s_ is empty.

Change-Id: Ib56a8a2513b0ae7ec8414b5ed3134e88b24dd2ac
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
Signed-off-by: AudioGod <audiogod@sonic-developers.com>
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>